# backend/openapi.json
{
  "components": {
    "schemas": {
      "ApiResponse_Any_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "ApiResponse[Any]",
        "type": "object"
      },
      "ApiResponse_DeleteMessageResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeleteMessageResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "ApiResponse[DeleteMessageResponse]",
        "type": "object"
      },
...
      "ApiResponse_NoneType_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data",
            "type": "null"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "ApiResponse[NoneType]",
        "type": "object"
      },
...
      "ApiResponse_RAGResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RAGResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "ApiResponse[RAGResponse]",
        "type": "object"
      },
      "ApiResponse_ReorderResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReorderResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "ApiResponse[ReorderResponse]",
        "type": "object"
      },
      "ApiResponse_SessionMessagesResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionMessagesResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "ApiResponse[SessionMessagesResponse]",
        "type": "object"
      },
      "ApiResponse_SessionResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "ApiResponse[SessionResponse]",
        "type": "object"
      },
...
      "ConfirmActionRequest": {
        "description": "高风险工具待确认动作的确认/取消请求模型",
        "properties": {
          "confirmed": {
            "default": true,
            "title": "Confirmed",
            "type": "boolean"
          },
          "pending_action_id": {
            "title": "Pending Action Id",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "required": [
          "pending_action_id"
        ],
        "title": "ConfirmActionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QueryRequest": {
        "description": "查询请求模型",
        "properties": {
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextSettings"
              },
              {
                "type": "null"
              }
            ]
          },
          "model_config_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Config Id"
          },
          "prompt_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Type"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "rag_retrieval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RagRetrievalSettings"
              },
              {
                "type": "null"
              }
            ]
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "skill_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Ids"
          },
          "tool_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Ids"
          }
        },
        "required": [
          "query"
        ],
        "title": "QueryRequest",
        "type": "object"
      },
      "RAGRequest": {
        "description": "RAG检索请求模型",
        "properties": {
          "query": {
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "RAGRequest",
        "type": "object"
      },
...
      "RegenerateRequest": {
        "description": "閲嶆柊鐢熸垚鍔╂墜鍥炵瓟璇锋眰妯″瀷",
        "properties": {
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextSettings"
              },
              {
                "type": "null"
              }
            ]
          },
          "model_config_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Config Id"
          },
          "prompt_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Type"
          },
          "rag_retrieval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RagRetrievalSettings"
              },
              {
                "type": "null"
              }
            ]
          },
          "skill_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Ids"
          },
          "tool_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Ids"
          }
        },
        "title": "RegenerateRequest",
        "type": "object"
      },
      "ReorderRequest": {
        "description": "重排序请求模型",
        "properties": {
          "documents": {
            "items": {
              "type": "string"
            },
            "title": "Documents",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query",
          "documents"
        ],
        "title": "ReorderRequest",
        "type": "object"
      },
...
      "SkillPayload": {
        "properties": {
          "always_on": {
            "default": false,
            "title": "Always On",
            "type": "boolean"
          },
          "default": {
            "default": true,
            "title": "Default",
            "type": "boolean"
          },
          "description": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "id": {
            "maxLength": 64,
            "minLength": 2,
            "title": "Id",
            "type": "string"
          },
          "instructions": {
            "default": "",
            "maxLength": 20000,
            "title": "Instructions",
            "type": "string"
          },
          "label": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "order": {
            "default": 100,
            "title": "Order",
            "type": "integer"
          },
          "routable": {
            "default": true,
            "title": "Routable",
            "type": "boolean"
          },
          "routing_examples": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Routing Examples",
            "type": "object"
          },
          "tools": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Tools",
            "type": "array"
          },
          "visibility": {
            "default": "public",
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "description"
        ],
        "title": "SkillPayload",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/chat/agent/confirm": {
      "post": {
        "description": "确认或取消一条高风险待确认动作，并以 SSE 流式返回执行结果。",
        "operationId": "confirm_agent_action_chat_agent_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "example": "data: {\"schema_version\":\"1.0\",\"type\":\"response\",\"content\":\"hello\"}\n\n",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Versioned Server-Sent Events stream"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Confirm Agent Action",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/agent/query/stream": {
      "post": {
        "description": "查询 Agent 流式响应。",
        "operationId": "query_stream_chat_agent_query_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "example": "data: {\"schema_version\":\"1.0\",\"type\":\"response\",\"content\":\"hello\"}\n\n",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Versioned Server-Sent Events stream"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Query Stream",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/chat/rag/query": {
      "post": {
        "description": "RAG 检索。",
        "operationId": "query_rag_chat_rag_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RAGRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RAGResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Query Rag",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/reorder": {
      "post": {
        "description": "对文档进行重排序。",
        "operationId": "reorder_documents_chat_reorder_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ReorderResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reorder Documents",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/session/{session_id}": {
      "delete": {
        "description": "删除会话。",
        "operationId": "delete_session_chat_session__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Session",
        "tags": [
          "chat"
        ]
      },
      "get": {
        "description": "获取会话信息。",
        "operationId": "get_session_chat_session__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SessionResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Session",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/session/{session_id}/messages": {
      "get": {
        "description": "获取会话原始消息列表（带消息 ID）。",
        "operationId": "get_session_messages_chat_session__session_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SessionMessagesResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Session Messages",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/session/{session_id}/messages/{message_id}": {
      "delete": {
        "description": "删除会话中的单条消息或相关消息。mode: single / pair / after。",
        "operationId": "delete_session_message_chat_session__session_id__messages__message_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "title": "Message Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "single",
              "title": "Mode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DeleteMessageResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Session Message",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/session/{session_id}/messages/{message_id}/regenerate/stream": {
      "post": {
        "description": "重新生成已有 assistant 消息，流式返回并覆盖原消息。",
        "operationId": "regenerate_session_message_stream_chat_session__session_id__messages__message_id__regenerate_stream_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "title": "Message Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "example": "data: {\"schema_version\":\"1.0\",\"type\":\"response\",\"content\":\"hello\"}\n\n",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Versioned Server-Sent Events stream"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Regenerate Session Message Stream",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/sessions": {
      "get": {
        "description": "获取当前用户所有会话。",
        "operationId": "get_all_sessions_chat_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Any_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get All Sessions",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/sessions/{user_id}": {
      "get": {
        "description": "获取用户所有会话 ID。",
        "operationId": "get_user_sessions_chat_sessions__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Any_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get User Sessions",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/skills": {
      "post": {
        "operationId": "create_skill_skills_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Any_"
                }
              }
            },
            "description": "Successful response using the canonical API envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Skill",
        "tags": [
          "skills"
        ]
      }
    },
...
  }
}